restructured to follow clean architecture principles
