+++ Data/Core/Abstracts/DataServiceAdapterAbstract.php
@@ -1,0 +1,8 @@
+<?php
+
+namespace Data\Core\Abstracts;
+
+abstract class DataServiceAdapterAbstract
+{
+    // Adapter abstract for data service apis.
+}
