+++ Data/Core/Exceptions/Types/DataException.php
@@ -1,0 +1,8 @@
+<?php
+
+namespace Data\Core\Exceptions\Types;
+
+class DataException extends \Exception
+{
+
+}
