+++ Data/Core/Exceptions/Types/DataServiceException.php
@@ -1,0 +1,8 @@
+<?php
+
+namespace Data\Core\Exceptions\Types;
+
+class DataServiceException extends \Exception
+{
+
+}
