+++ Data/Models/Hello.php
@@ -1,0 +1,11 @@
+<?php
+
+namespace Data\Models;
+
+class Hello
+{
+	const TABLE = 'hello';
+
+	const STATUS_ACTIVE = 'A';
+	const STATUS_INACTIVE = 'D';
+}
