+++ Data/Models/User.php
@@ -1,0 +1,11 @@
+<?php
+
+namespace Data\Models;
+
+class User
+{
+	const TABLE = 'user';
+
+	const STATUS_ACTIVE = 'A';
+	const STATUS_INACTIVE = 'D';
+}
