+++ Domain/Core/Abstracts/LogicServiceAdapterAbstract.php
@@ -1,0 +1,8 @@
+<?php
+
+namespace Domain\Core\Abstracts;
+
+abstract class LogicServiceAdapterAbstract
+{
+    // Adapter abstract for domain logic level service apis.
+}
