+++ Domain/Core/Adapters/Service/Example/ExampleLogicServiceAdapter.php
@@ -1,0 +1,10 @@
+<?php
+
+namespace Domain\Core\Adapters\Service\Example;
+
+use Domain\Core\Abstracts\LogicServiceAdapterAbstract;
+
+class ExampleLogicServiceAdapter
+{
+    // Adapter abstract for domain logic level service apis.
+}
