+++ Domain/Core/Exceptions/Types/LogicServiceException.php
@@ -1,0 +1,8 @@
+<?php
+
+namespace Domain\Core\Exceptions\Types;
+
+class LogicServiceException extends \Exception
+{
+
+}
