+++ Domain/Core/Exceptions/Types/LogicalException.php
@@ -1,0 +1,8 @@
+<?php
+
+namespace Domain\Core\Exceptions\Types;
+
+class LogicalException extends \Exception
+{
+
+}
