--- api/Core/Abstracts/ServiceAdapterAbstract.php
@@ -1,8 +1,0 @@
-<?php
-
-namespace Api\Core\Abstracts;
-
-abstract class ServiceAdapterAbstract
-{
-    // Adapter abstract for data service apis.
-}
