--- api/Core/Adapters/Service/ExampleServiceAdapter.php
@@ -1,10 +1,0 @@
-<?php
-
-namespace Api\Core\Adapters\Service;
-
-use Api\Core\Abstracts\ServiceAdapterAbstract;
-
-class ExampleServiceAdapter
-{
-    // Adapter abstract for data service apis.
-}
