--- api/Core/Exceptions/Types/DataException.php
@@ -1,8 +1,0 @@
-<?php
-
-namespace Api\Core\Exceptions\Types;
-
-class DataException extends \Exception
-{
-
-}
