--- api/Core/Exceptions/Types/LogicalException.php
@@ -1,8 +1,0 @@
-<?php
-
-namespace Api\Core\Exceptions\Types;
-
-class LogicalException extends \Exception
-{
-
-}
