--- api/Core/Exceptions/Types/ServiceException.php
@@ -1,8 +1,0 @@
-<?php
-
-namespace Api\Core\Exceptions\Types;
-
-class ServiceException extends \Exception
-{
-
-}
