--- api/Data/Models/Hello.php
@@ -1,11 +1,0 @@
-<?php
-
-namespace Api\Data\Models;
-
-class Hello
-{
-	const TABLE = 'hello';
-
-	const STATUS_ACTIVE = 'A';
-	const STATUS_INACTIVE = 'D';
-}
