--- api/Data/Models/User.php
@@ -1,11 +1,0 @@
-<?php
-
-namespace Api\Data\Models;
-
-class User
-{
-	const TABLE = 'user';
-
-	const STATUS_ACTIVE = 'A';
-	const STATUS_INACTIVE = 'D';
-}
