--- api/Logic/Hello/HelloException.php
@@ -1,21 +1,0 @@
-<?php
-
-namespace Api\Logic\Hello;
-
-use Api\Core\Exceptions\Types\LogicalException;
-
-class HelloException extends LogicalException
-{
-    const BASE_ERROR_CODE = 1000;
-
-    public static function unknown()
-    {
-        throw new HelloException("Unknown Hello Exception", self::BASE_ERROR_CODE);
-    }
-
-    public static function exception()
-    {
-        throw new HelloException("Hello Exception", self::BASE_ERROR_CODE + 1);
-    }
-
-}
